Extract errorsToFlashbarItems into its own file
